requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Flask Frontend
Flask==3.0.0
Jinja2==3.1.2

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
ruff==0.1.6
black==23.11.0

# Logging
structlog==23.2.0
